Version bump, remove old and patches. It's likely not going to work anyway, since pcsc-lite has changed on its own, but at least provide what upstream provides.
(Portage version: 2.2.0_alpha10/cvs/Linux x86_64)

+++ sys-apps/ifd-gempc/ifd-gempc-1.0.6.ebuild
@@ -1,8 +1,8 @@
-# Copyright 1999-2009 Gentoo Foundation
+# Copyright 1999-2010 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-apps/ifd-gempc/Attic/ifd-gempc-1.0.4.ebuild,v 1.2 2009/05/16 09:23:13 robbat2 Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-apps/ifd-gempc/ifd-gempc-1.0.6.ebuild,v 1.1 2010/12/26 18:35:31 flameeyes Exp $
 
-inherit eutils toolchain-funcs
+inherit toolchain-funcs
 
 DESCRIPTION="GemCore based PC/SC reader drivers for pcsc-lite"
 HOMEPAGE="http://ludovic.rousseau.free.fr/softwares/ifd-GemPC"
@@ -16,13 +16,6 @@ RDEPEND=">=sys-apps/pcsc-lite-1.2.9_beta7
 DEPEND="${RDEPEND}
 	dev-util/pkgconfig"
 
-src_unpack() {
-	unpack ${A}
-	cd "${S}"
-
-	epatch "${FILESDIR}/${P}-respect_LDFLAGS.patch"
-}
-
 src_compile() {
 	emake CC="$(tc-getCC)" || die "emake failed"
 }
@@ -30,17 +23,19 @@ src_compile() {
 src_install () {
 	local pcscdir="$(pkg-config --variable=usbdropdir libpcsclite)"
 	local conf="/etc/reader.conf.d/${PN}.conf"
-	emake install CC="$(tc-getCC)" DESTDIR="${D}" || die
+	emake CC="$(tc-getCC)" DESTDIR="${D}" install || die "emake install failed"
 	dodoc README*
 	dodir "$(dirname "${conf}")"
 	insinto "$(dirname "${conf}")"
 	newins "${FILESDIR}/reader.conf" "$(basename "${conf}")"
 	sed -e "s:%PCSC_DRIVERS_DIR%:${pcscdir}:g" -e "s:%libGemPC410%:libGemPC410.so.${PV}:g" -i "${D}${conf}"
+}
 
-	einfo "NOTICE:"
-	einfo "1. if you are using GemPC410 modify ${conf}"
-	einfo "2. run update-reader.conf, yes this is a command..."
-	einfo "3. restart pcscd"
+pkg_postinst() {
+	elog "NOTICE:"
+	elog "1. If you are using GemPC410, modify ${conf}"
+	elog "2. Run update-reader.conf, yes this is a command..."
+	elog "3. Restart pcscd"
 }
 
 pkg_postrm() {
@@ -48,10 +43,10 @@ pkg_postrm() {
 	# Without this, pcscd will not start next time.
 	#
 	local conf="/etc/reader.conf.d/${PN}.conf"
-	if ! [ -f "$(grep LIBPATH "${conf}" | sed 's/LIBPATH *//' | sed 's/ *$//g' | head -n 1)" ]; then
-		rm "${conf}"
+	if ! [[ -f "${conf}" && -f "$(grep LIBPATH "${conf}" | sed 's/LIBPATH *//' | sed 's/ *$//g' | head -n 1)" ]]; then
+		[[ -f "${conf}" ]] && rm "${conf}"
 		update-reader.conf
-		einfo "NOTICE:"
-		einfo "You need to restart pcscd"
+		elog "NOTICE:"
+		elog "You need to restart pcscd"
 	fi
 }
